+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 ## 1.77 (unreleased)
 
-
+-   Feat [UploadTasks]: Add the option to deliver XMLs based on files.
 -   Feat [ObjectImport]: Made it possible to create individual imports through objectimporttasks
 -   Enhancement [ImportTasks]: Convert to celery tasks, for faster handling
 -   Bugfix [Import]: The most recent tube top was shown within the GMW_Construction, instead of the original.